+++ 102/HW_2/HW_2/Controllers/MarketViewController.swift
@@ -1,0 +1,17 @@
+//
+//  MarketViewController.swift
+//  HW_2
+//
+//  Created by Тимур Хайруллин on 18.09.2023.
+//
+
+import UIKit
+
+class MarketViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        self.navigationItem.setHidesBackButton(true, animated: false)
+    }
+}
